moving ChatModels to chat_api
